Refactor import path for DOMContent to use centralized components location
