+++ components/rewards/dom-content.tsx
